Extend locking
Another random plotting error occured in a
bg process without any helpful traceback.

All calls to savefig should now be covered with locks.

Change lock name to 'plotting' everywhere.
That may be over-the-top, but it doesn't
have a noticable effect on performance and it
should show if locking is even a fully working solution.