trainer: Don't adjust raw intensities in raw PNG
plt.imsave() without explicit vmin and vmax parameters adjust
intensities so the minimum and maximum values are at 0 and 1,
respectively. This is usually not wanted and can cause misleading
visualizations, so the actual intensity values are now written to the
"raw" preview png instead.

Theoretically we have the same issue with prediction outputs, but after
experimenting with the same "fix" (vmin=0, vmax=1) on predictions, it looks
like we should keep that as it is right now, because the limited value range
(~ between 0.4 and 0.6) would make predictions very hard to understand
otherwise.